+++ popuko/slack/__init__.py
@@ -2,7 +2,7 @@ import json
 
 import requests
 
-from jewelpet.conf import settings
+from popuko.conf import settings
 
 
 def post(message):
